pyproject.toml
[project]
name = "langgraph-agentcore-sample"
version = "0.1.0"
description = "LangGraph agent with web search deployed to AWS Bedrock AgentCore"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
# Runtime dependencies only - these are what get packaged in the Docker container
dependencies = [
    # Core LangChain/LangGraph
    "langchain~=1.2.0",
    "langgraph~=1.0.5",
    "langchain-aws~=1.1.0",
    "langchain-community~=0.4.1",
    "langchain-core~=1.2.4",
    # Tools and utilities
    "langchain-tavily>=0.2.0",
    "python-dotenv~=1.2.1",
    # Retry logic
    "tenacity~=9.0.0",
    # AWS Bedrock AgentCore runtime SDK
    "bedrock-agentcore~=1.1.1",
    # OpenTelemetry instrumentation
    "opentelemetry-instrumentation-langchain~=0.45.6",
    "aws-opentelemetry-distro~=0.12.2",
]

[project.optional-dependencies]
# Local testing - NOT included in Docker container
local = [
    "bedrock-agentcore-starter-toolkit>=0.2.0",
]
# Deployment tools - NOT included in Docker container
deploy = [
    "aws-cdk-lib~=2.233.0",                       # CDK for infrastructure
    "constructs~=10.4.4",                         # CDK dependency
    "typer>=0.16.0",                               # CLI framework with Rich
    "websocket-client~=1.8.0",                    # WebSocket for agent invocation
]
# NiceGUI testing UI
ui = [
    "nicegui>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest~=9.0.2",
    "pytest-cov~=7.0.0",
    "ruff>=0.8.0",
]

[tool.setuptools]
py-modules = ["langgraph_agent_web_search"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
# E402: Module level import not at top of file
# This is intentional for agentcore apps - the import must be after function definitions
"langgraph_agent_web_search.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
